STY: moved DialogLoadMask to a separate file
